requirements.txt
crewai
pyYAML
scikit-learn
dotenv
langchain_community
pydantic
duckduckgo-search
streamlit